requirements.txt
fastapi
requests
google-auth